enable on-device builds
Partial compatibility for on-device builds.

There is no guarantee that it will be possible to build all available
packages and built packages will have same reliability that cross-compiled
but should solve "self-hosting" problems as much as possible.

+++ scripts/build/setup/termux_setup_golang.sh
@@ -19,20 +19,35 @@ termux_setup_golang() {
 		termux_error_exit "Unsupported arch: $TERMUX_ARCH"
 	fi
 
-	local TERMUX_GO_VERSION=go1.12.7
-	local TERMUX_GO_PLATFORM=linux-amd64
+	if [ -z "$TERMUX_ON_DEVICE_BUILD" ]; then
+		local TERMUX_GO_VERSION=go1.12.7
+		local TERMUX_GO_PLATFORM=linux-amd64
 
-	local TERMUX_BUILDGO_FOLDER=$TERMUX_COMMON_CACHEDIR/${TERMUX_GO_VERSION}
-	export GOROOT=$TERMUX_BUILDGO_FOLDER
-	export PATH=$GOROOT/bin:$PATH
+		local TERMUX_BUILDGO_FOLDER=$TERMUX_COMMON_CACHEDIR/${TERMUX_GO_VERSION}
+		export GOROOT=$TERMUX_BUILDGO_FOLDER
+		export PATH=$GOROOT/bin:$PATH
 
-	if [ -d "$TERMUX_BUILDGO_FOLDER" ]; then return; fi
+		if [ -d "$TERMUX_BUILDGO_FOLDER" ]; then return; fi
 
-	local TERMUX_BUILDGO_TAR=$TERMUX_COMMON_CACHEDIR/${TERMUX_GO_VERSION}.${TERMUX_GO_PLATFORM}.tar.gz
-	rm -Rf "$TERMUX_COMMON_CACHEDIR/go" "$TERMUX_BUILDGO_FOLDER"
-	termux_download https://storage.googleapis.com/golang/${TERMUX_GO_VERSION}.${TERMUX_GO_PLATFORM}.tar.gz \
-		"$TERMUX_BUILDGO_TAR" \
-		66d83bfb5a9ede000e33c6579a91a29e6b101829ad41fffb5c5bb6c900e109d9
+		local TERMUX_BUILDGO_TAR=$TERMUX_COMMON_CACHEDIR/${TERMUX_GO_VERSION}.${TERMUX_GO_PLATFORM}.tar.gz
+		rm -Rf "$TERMUX_COMMON_CACHEDIR/go" "$TERMUX_BUILDGO_FOLDER"
+		termux_download https://storage.googleapis.com/golang/${TERMUX_GO_VERSION}.${TERMUX_GO_PLATFORM}.tar.gz \
+			"$TERMUX_BUILDGO_TAR" \
+			66d83bfb5a9ede000e33c6579a91a29e6b101829ad41fffb5c5bb6c900e109d9
 
-	( cd "$TERMUX_COMMON_CACHEDIR"; tar xf "$TERMUX_BUILDGO_TAR"; mv go "$TERMUX_BUILDGO_FOLDER"; rm "$TERMUX_BUILDGO_TAR" )
+		( cd "$TERMUX_COMMON_CACHEDIR"; tar xf "$TERMUX_BUILDGO_TAR"; mv go "$TERMUX_BUILDGO_FOLDER"; rm "$TERMUX_BUILDGO_TAR" )
+	else
+		if [ "$(dpkg-query -W -f '${db:Status-Status}\n' golang 2>/dev/null)" != "installed" ]; then
+			echo "Package 'golang' is not installed."
+			echo "You can install it with"
+			echo
+			echo "  pkg install golang"
+			echo
+			echo "or build it from source with"
+			echo
+			echo "  ./build-package.sh golang"
+			echo
+			exit 1
+		fi
+	fi
 }
